pyproject.toml
[tool.poetry]
name = "certitude"
version = "0.3.1"
description = "Determine if a URL may be malicious"
authors = ["Paul Wiper <paul.wiper@tno.nl>", "Marie Beth van Egmond <marie_beth.vanegmond@tno.nl"]
license = "MPL-2.0"
include = ["certitude/utils/data/*"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.1"
requests = "^2.25.1"
scikit-learn = "^0.24.1"
whois = "^0.9.13"
joblib = "^1.0.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
pre-commit = "^2.12.1"
pylic = "^1.2.1"
bandit = "^1.7.0"
mypy = "^0.812"
flake8 = "^3.9.2"
build = "^0.4.0"
pytest = "^6.2.4"
vulture = "^2.3"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylic]
safe_licenses = [
    "Apache Software License",
    "MIT License",
    "BSD License",
]
